requirements.txt
asgiref==3.2.10
click==7.1.2
Django==3.0.8
djangorestframework==3.11.0
imutils==0.5.3
joblib==0.16.0
nltk==3.5
numpy==1.19.0
opencv-python==4.3.0.36
Pillow==7.2.0
pytz==2020.1
regex==2020.6.8
sqlparse==0.3.1
tesseract==0.1.3
tqdm==4.47.0
urllib3==1.25.9
